app/models/lstm_model.py
"""
LSTM model implementation for Trading AI System
Deep learning model for sequence prediction
"""
import numpy as np
import pandas as pd
import logging
import torch
import torch.nn as nn
from sklearn.preprocessing import MinMaxScaler

logger = logging.getLogger(__name__)

class LSTMTradingModel(nn.Module):
    """LSTM model for trading sequence prediction"""

    def __init__(self, input_size=10, hidden_size=50, num_layers=2, dropout=0.2):
        super(LSTMTradingModel, self).__init__()
        self.hidden_size = hidden_size
        self.num_layers = num_layers

        self.lstm = nn.LSTM(
            input_size, hidden_size, num_layers,
            batch_first=True, dropout=dropout
        )
        self.fc = nn.Linear(hidden_size, 1)
        self.sigmoid = nn.Sigmoid()
        self.dropout = nn.Dropout(dropout)

    def forward(self, x):
        lstm_out, _ = self.lstm(x)
        last_output = lstm_out[:, -1, :]
        dropped = self.dropout(last_output)
        output = self.fc(dropped)
        return self.sigmoid(output)

class LSTMTradingWrapper:
    """Wrapper for LSTM trading model"""

    def __init__(self, config):
        self.config = config
        self.model = None
        self.scaler = MinMaxScaler()
        self.is_trained = False
        self.sequence_length = 60  # Look back 60 time steps

    def train(self, training_data):
        """Train LSTM model (placeholder)"""
        logger.info("🧠 LSTM training not fully implemented in this demo")
        logger.info("📝 Use XGBoost model for full functionality")
        return False

    def predict_single(self, features_dict):
        """Placeholder prediction"""
        return {'prediction': 1, 'probability': 0.5}