requirements.txt
# Core Dependencies
flask==2.3.3
flask-socketio==5.3.6
flask-sqlalchemy==3.0.5
gunicorn==21.2.0

# Data Processing & ML
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.1.0
torch==2.0.1
joblib==1.3.2

# Technical Analysis
ta==0.10.2

# Database
sqlalchemy==2.0.21
alembic==1.12.0

# Networking & API
requests==2.31.0
websocket-client==1.6.3
python-binance==1.0.19

# Scheduling & Background Tasks
APScheduler==3.10.4

# Logging & Monitoring
structlog==23.1.0
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Production
eventlet==0.33.3
python-dotenv==1.0.0